scripts/injector/requirements.txt
argparse >= 1.4.0
pymongo >= 4.3.3
psycopg >= 3.1.8
cassandra-driver >= 3.25.0
tqdm >= 4.65.0
